Introduce organization_primitives to break dependency cycle
