+++ front/lib/api/workos/organization.ts
@@ -10,6 +10,7 @@ import uniqueId from "lodash/uniqueId";
 import { config } from "@app/lib/api/regions/config";
 import { getWorkOS } from "@app/lib/api/workos/client";
 import { invalidateWorkOSOrganizationsCacheForUserId } from "@app/lib/api/workos/organization_membership";
+import { getWorkOSOrganization } from "@app/lib/api/workos/organization_primitives";
 import { MembershipResource } from "@app/lib/resources/membership_resource";
 import { WorkspaceResource } from "@app/lib/resources/workspace_resource";
 import { WorkOSPortalIntent } from "@app/lib/types/workos";
@@ -17,35 +18,6 @@ import { concurrentExecutor } from "@app/lib/utils/async_utils";
 import logger from "@app/logger/logger";
 import type { LightWorkspaceType, Result } from "@app/types";
 import { Err, normalizeError, Ok } from "@app/types";
-
-function isWorkOSNotFoundEntityError(error: unknown): boolean {
-  return (
-    error instanceof Error &&
-    "status" in error &&
-    error.status === 404 &&
-    "code" in error &&
-    error.code === "entity_not_found"
-  );
-}
-
-export async function getWorkOSOrganization(
-  workspace: LightWorkspaceType
-): Promise<Result<Organization | undefined, Error>> {
-  try {
-    const result = await getWorkOS().organizations.getOrganizationByExternalId(
-      workspace.sId
-    );
-
-    return new Ok(result);
-  } catch (error) {
-    // If the organization is not found, return undefined.
-    if (isWorkOSNotFoundEntityError(error)) {
-      return new Ok(undefined);
-    }
-
-    return new Err(new Error("Failed to get WorkOS organization."));
-  }
-}
 
 export async function getOrCreateWorkOSOrganization(
   workspace: LightWorkspaceType,
@@ -162,57 +134,6 @@ export async function addWorkOSOrganizationDomain(
   return new Ok(undefined);
 }
 
-export async function removeWorkOSOrganizationDomain(
-  workspace: LightWorkspaceType,
-  { domain }: { domain: string }
-): Promise<Result<void, Error>> {
-  const organizationRes = await getWorkOSOrganization(workspace);
-  if (organizationRes.isErr()) {
-    return new Err(organizationRes.error);
-  }
-
-  const organization = organizationRes.value;
-  if (!organization) {
-    return new Err(
-      new Error("WorkOS organization not found for this workspace.")
-    );
-  }
-
-  return removeWorkOSOrganizationDomainFromOrganization(organization, {
-    domain,
-  });
-}
-
-export async function removeWorkOSOrganizationDomainFromOrganization(
-  organization: Organization,
-  { domain }: { domain: string }
-): Promise<Result<void, Error>> {
-  await getWorkOS().organizations.updateOrganization({
-    organization: organization.id,
-    domainData: organization.domains
-      .filter(
-        (d) =>
-          d.domain !== domain && d.state === OrganizationDomainState.Verified
-      )
-      .map((d) => ({
-        domain: d.domain,
-        state: DomainDataState.Verified,
-      })),
-  });
-
-  // WARN: Hacky update done after the domain data, so that it trigger
-  // the webhook. Should be remove once WorkOS send us webhook when just
-  // the domains change.
-  await getWorkOS().organizations.updateOrganization({
-    organization: organization.id,
-    metadata: {
-      _webhookTrigger: uniqueId(),
-    },
-  });
-
-  return new Ok(undefined);
-}
-
 export async function updateWorkOSOrganizationName(
   workspace: LightWorkspaceType
 ): Promise<Result<void, Error>> {
@@ -248,18 +169,6 @@ export async function updateWorkOSOrganizationName(
   }
 
   return new Ok(undefined);
-}
-
-export async function listWorkOSOrganizationsWithDomain(
-  domain: string
-): Promise<Organization[]> {
-  const workOS = getWorkOS();
-  const organizations = await workOS.organizations.listOrganizations({
-    domains: [domain],
-    limit: 100,
-  });
-
-  return organizations.data;
 }
 
 // Mapping WorkOSPortalIntent to GeneratePortalLinkIntent,
