+++ front/lib/resources/workspace_resource.ts
@@ -6,7 +6,7 @@ import {
   listWorkOSOrganizationsWithDomain,
   removeWorkOSOrganizationDomain,
   removeWorkOSOrganizationDomainFromOrganization,
-} from "@app/lib/api/workos/organization";
+} from "@app/lib/api/workos/organization_primitives";
 import type { Authenticator } from "@app/lib/auth";
 import type { ResourceLogJSON } from "@app/lib/resources/base_resource";
 import { BaseResource } from "@app/lib/resources/base_resource";
